Restructured classes into city.py and attractions.py
